Separate test environment for NullableRefTypes
