+++ Tests/RewritingMethods.cs
@@ -78,37 +78,10 @@ public class RewritingMethods :
     }
 
     [Fact]
-    public Task RequiresNonNullArgumentWhenNullableReferenceTypeNotUsed()
-    {
-        var type = AssemblyWeaver.Assembly.GetType(nameof(NullableReferenceTypeClass));
-        var sample = (dynamic)Activator.CreateInstance(type);
-        var exception = Assert.Throws<ArgumentNullException>(() => { sample.SomeMethod(null, ""); });
-        Assert.Equal("nonNullArg", exception.ParamName);
-        return Verify(exception.Message);
-    }
-
-    [Fact]
-    public void AllowsNullWhenNullableReferenceTypeUsed()
-    {
-        var type = AssemblyWeaver.Assembly.GetType(nameof(NullableReferenceTypeClass));
-        var sample = (dynamic)Activator.CreateInstance(type);
-        sample.SomeMethod("", null);
-    }
-
-    [Fact]
     public Task RequiresNonNullMethodReturnValue()
     {
         var type = AssemblyWeaver.Assembly.GetType("SimpleClass");
         var sample = (dynamic) Activator.CreateInstance(type);
-        var exception = Assert.Throws<InvalidOperationException>(() => sample.MethodWithReturnValue(true));
-        return Verify(exception.Message);
-    }
-
-    [Fact]
-    public Task RequiresNonNullMethodReturnValueWhenNullableReferenceTypeNotUsed()
-    {
-        var type = AssemblyWeaver.Assembly.GetType(nameof(NullableReferenceTypeClass));
-        var sample = (dynamic)Activator.CreateInstance(type);
         var exception = Assert.Throws<InvalidOperationException>(() => sample.MethodWithReturnValue(true));
         return Verify(exception.Message);
     }
@@ -127,14 +100,6 @@ public class RewritingMethods :
     {
         var type = AssemblyWeaver.Assembly.GetType("SimpleClass");
         var sample = (dynamic) Activator.CreateInstance(type);
-        sample.MethodAllowsNullReturnValue();
-    }
-
-    [Fact]
-    public void AllowsNullReturnValueWhenNullableReferenceType()
-    {
-        var type = AssemblyWeaver.Assembly.GetType(nameof(NullableReferenceTypeClass));
-        var sample = (dynamic)Activator.CreateInstance(type);
         sample.MethodAllowsNullReturnValue();
     }
