+++ TestsNullableRefTypes/TestsNullableRefTypes.csproj
@@ -1,0 +1,19 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <TargetFrameworks>net472;netcoreapp2.2</TargetFrameworks>
+    <DisableFody>true</DisableFody>
+  </PropertyGroup>
+  <ItemGroup>
+    <PackageReference Include="FodyHelpers" Version="6.0.6" />
+    <PackageReference Include="xunit" Version="2.4.1" />
+    <PackageReference Include="XunitContext" Version="1.6.0" />
+    <PackageReference Include="Verify.Xunit" Version="1.8.0" />
+    <PackageReference Include="xunit.runner.visualstudio" Version="2.4.1" />
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.4.0" />
+    <Reference Include="Microsoft.CSharp" />
+    <ProjectReference Include="..\AssemblyToProcessNullableRefTypes\AssemblyToProcessNullableRefTypes.csproj" />
+    <ProjectReference Include="..\NetFrameworkSmokeTest\NetFrameworkSmokeTest.csproj" />
+    <ProjectReference Include="..\NullGuard.Fody\NullGuard.Fody.csproj" />
+    <ProjectReference Include="..\NullGuard\NullGuard.csproj" />
+  </ItemGroup>
+</Project>
